Update dependencies, fixes build errors and removes binding

+++ app/src/main/java/io/sikorka/android/core/accounts/PassphraseValidator.kt
@@ -3,5 +3,5 @@ package io.sikorka.android.core.accounts
 
 interface PassphraseValidator {
   @ValidationResult.Code
-  fun validate(passphrase: String, passphraseConfirmation: String): Long
+  fun validate(passphrase: String, passphraseConfirmation: String): Int
 }
