+++ app/src/main/java/io/sikorka/android/ui/accounts/account_creation/AccountCreationDialog.kt
@@ -80,7 +80,7 @@ class AccountCreationDialog : DialogFragment(), AccountCreationDialogView {
     presenter.attach(this)
   }
 
-  override fun showError(@Code code: Long) {
+  override fun showError(@Code code: Int) {
     clearErrors()
     when (code) {
       ValidationResult.CONFIRMATION_MISMATCH -> {
